fix(migration): close out remaining PR 169 PR review gaps
- tasks/task_lists.index widened to a 64-bit column and creation/update index
  validation relaxed to signed integers so the frontend's -Date.now() top-insert
  scheme (and imported Instant data) no longer overflows/rejects
- enforce compatible migration flags: IKUYO_BACKEND_AUTH now requires
  IKUYO_BACKEND_TRIP_READS to avoid a signed-in-but-empty InstantDB read state
- mount usePeriodicTripSync on the trip page (backend-read mode) so collaborator
  edits are reflected via sync polling; add a refreshTrip store action and refactor
  the fetch+merge path into a shared helper with unmount abort
- keep the sync hook's interval stable across renders (callback held in a ref)
- stream JSONL reads in the import command (generator) instead of materializing
  each file into memory, and count via iterator_count for verification

+++ backend/database/migrations/2026_01_01_000003_create_comments_and_tasks_tables.php
@@ -12,7 +12,9 @@ return new class extends Migration
             $table->string('id', 40)->primary();
             $table->string('trip_id', 40);
             $table->string('title');
-            $table->integer('index');
+            // 64-bit so negative `-Date.now()` insertion indexes (from the
+            // frontend / Instant imports) fit; 32-bit would overflow.
+            $table->bigInteger('index');
             $table->tinyInteger('status'); // open / done
             $table->bigInteger('created_at_ms');
             $table->bigInteger('updated_at_ms');
@@ -22,7 +24,7 @@ return new class extends Migration
         Schema::create('tasks', function (Blueprint $table): void {
             $table->string('id', 40)->primary();
             $table->string('task_list_id', 40);
-            $table->integer('index');
+            $table->bigInteger('index');
             $table->string('title');
             $table->text('description')->nullable();
             $table->tinyInteger('status'); // 0 todo / 1 done
